requirements.txt
anyascii==0.3.3
asgiref==3.9.1
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
croniter==6.0.0
cryptography==45.0.6
defusedxml==0.7.1
Django==5.2.4
django-allauth==65.10.0
django-axes==8.0.0
django-filter==25.1
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.2.2
django-taggit==6.1.0
django-tasks==0.8.1
django-treebeard==4.7.1
djangorestframework==3.16.1
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
git-filter-repo==2.47.0
idna==3.10
laces==0.1.2
msal==1.33.0
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
pillow_heif==1.1.0
pip-check==3.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
telepath==0.3.1
terminaltables==3.1.10
typing_extensions==4.14.1
urllib3==2.5.0
wagtail==7.1
Willow==1.11.0
